+++ src/leapfrogai_api/routers/openai/requests/create_modify_assistant_request.py
@@ -1,20 +1,11 @@
 from __future__ import annotations
 
 import logging
-
 from contextlib import suppress
 from fastapi import HTTPException, status
 from pydantic import BaseModel, Field
 from typing import Literal
 
-from leapfrogai_api.backend.rag.index import IndexingService
-from leapfrogai_api.backend.types import CreateVectorStoreRequest
-from leapfrogai_api.data.crud_vector_store import CRUDVectorStore, FilterVectorStore
-from leapfrogai_api.routers.supabase_session import Session
-from leapfrogai_api.utils.validate_tools import (
-    validate_assistant_tool,
-    validate_tool_resources,
-)
 from openai.types.beta import AssistantTool
 from openai.types.beta.assistant import (
     ToolResources as BetaAssistantToolResources,
@@ -23,6 +14,16 @@ from openai.types.beta.assistant import (
 from openai.types.beta.assistant_tool import FileSearchTool
 
 logger = logging.getLogger(__name__)
+
+
+from leapfrogai_api.backend.rag.index import IndexingService
+from leapfrogai_api.typedef.vectorstores import CreateVectorStoreRequest
+from leapfrogai_api.data.crud_vector_store import CRUDVectorStore, FilterVectorStore
+from leapfrogai_api.routers.supabase_session import Session
+from leapfrogai_api.utils.validate_tools import (
+    validate_assistant_tool,
+    validate_tool_resources,
+)
 
 
 class CreateAssistantRequest(BaseModel):
